ClimbingMotor class
Migrate climbing code from PlayerMotor to new class, ClimbingMotor.
