+++ Assets/Scripts/Game/Player/ClimbingMotor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 727122ccc0158b8439805f6f3b64b449
+timeCreated: 1525739609
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
