+++ Assets/Scripts/Game/PlayerMotor.cs
@@ -98,13 +98,8 @@ namespace DaggerfallWorkshop.Game
         private bool jumping = false;
         private bool standingStill = false;
 
-        private bool failedClimbingCheck = false;
-        private bool isClimbing = false;
-        private float climbingStartTimer = 0;
-        private float climbingContinueTimer = 0;
-        private uint timeOfLastClimbingCheck = 0;
-        private bool showClimbingModeMessage = true;
-        private Vector2 lastHorizontalPosition = Vector2.zero;
+
+        private ClimbingMotor climbingMotor;
         private PlayerHeightChanger heightChanger;
         private PlayerSpeedChanger speedChanger;
 
@@ -145,12 +140,6 @@ namespace DaggerfallWorkshop.Game
         {
             get { return isRiding; }
             set { isRiding = value; }
-        }
-
-        public bool IsClimbing
-        {
-            get { return isClimbing; }
-            set { isClimbing = value; }
         }
 
         public bool IsMovingLessThanHalfSpeed
@@ -219,6 +208,7 @@ namespace DaggerfallWorkshop.Game
             slideLimit = controller.slopeLimit - .1f;
             jumpTimer = antiBunnyHopFactor;
             //mainCamera = GameManager.Instance.MainCamera;
+            climbingMotor = GetComponent<ClimbingMotor>();
             heightChanger = GetComponent<PlayerHeightChanger>();
             levitateMotor = GetComponent<LevitateMotor>();
 
@@ -251,68 +241,15 @@ namespace DaggerfallWorkshop.Game
                 return;
             }
 
-            if (isClimbing)
-                collisionFlags = CollisionFlags.Sides;
-            // Get collision flags for swimming as well, so it's possible to climb out of water TODO: Collision flags from swimming aren't working
-            else if (levitateMotor.IsSwimming)
-                collisionFlags = levitateMotor.CollisionFlags;
+            climbingMotor.ClimbingCheck(ref collisionFlags);
 
-            // Climbing
-            uint gameMinutes = DaggerfallUnity.Instance.WorldTime.DaggerfallDateTime.ToClassicDaggerfallTime();
-            if (!InputManager.Instance.HasAction(InputManager.Actions.MoveForwards)
-                || (collisionFlags & CollisionFlags.Sides) == 0
-                || failedClimbingCheck
-                || levitateMotor.IsLevitating
-                || isRiding
-                || Vector2.Distance(lastHorizontalPosition, new Vector2(controller.transform.position.x, controller.transform.position.z)) >= (0.003f)) // Approximation based on observing classic in-game
-            {
-                isClimbing = false;
-                showClimbingModeMessage = true;
-                climbingStartTimer = 0;
-                timeOfLastClimbingCheck = gameMinutes;
-            }
-            else
-            {
-                if (climbingStartTimer <= (systemTimerUpdatesPerSecond * 14))
-                    climbingStartTimer += Time.deltaTime;
-                else
-                {
-                    if (!isClimbing)
-                    {
-                        if (showClimbingModeMessage)
-                            DaggerfallUI.AddHUDText(UserInterfaceWindows.HardStrings.climbingMode);
-                        // Disable further showing of climbing mode message until current climb attempt is stopped
-                        // to keep it from filling message log
-                        showClimbingModeMessage = false;
-                        isClimbing = true;
-                    }
-
-                    // Initial check to start climbing
-                    if ((gameMinutes - timeOfLastClimbingCheck) > 18)
-                    {
-                        Entity.PlayerEntity player = GameManager.Instance.PlayerEntity;
-                        player.TallySkill(DFCareer.Skills.Climbing, 1);
-                        timeOfLastClimbingCheck = gameMinutes;
-                        if (UnityEngine.Random.Range(1, 101) > 95)
-                        {
-                            if (UnityEngine.Random.Range(1, 101) > player.Skills.GetLiveSkillValue(DFCareer.Skills.Climbing))
-                            {
-                                isClimbing = false;
-                                failedClimbingCheck = true;
-                            }
-                        }
-                    }
-                }
-            }
-
-            if (isClimbing)
+            if (climbingMotor.IsClimbing)
             {
                 falling = false;
-                ClimbMovement();
             }
 
             // Do nothing if player levitating/swimming or climbing - replacement motor will take over movement for levitating/swimming
-            if (levitateMotor && (levitateMotor.IsLevitating || levitateMotor.IsSwimming) || isClimbing)
+            if (levitateMotor && (levitateMotor.IsLevitating || levitateMotor.IsSwimming) || climbingMotor.IsClimbing)
                 return;
 
             //float inputX = Input.GetAxis("Horizontal");
@@ -495,9 +432,6 @@ namespace DaggerfallWorkshop.Game
             // Move the controller, and set grounded true or false depending on whether we're standing on something
             collisionFlags = controller.Move(moveDirection * Time.deltaTime);
 
-            // Get pre-movement position for climbing check
-            lastHorizontalPosition = new Vector2(controller.transform.position.x, controller.transform.position.z);
-
             grounded = (collisionFlags & CollisionFlags.Below) != 0;
 
             // Moving platforms support
@@ -509,29 +443,6 @@ namespace DaggerfallWorkshop.Game
                 // If you want to support moving platform rotation as well:
                 activeGlobalPlatformRotation = transform.rotation;
                 activeLocalPlatformRotation = Quaternion.Inverse(activePlatform.rotation) * transform.rotation;
-            }
-        }
-
-        private void ClimbMovement()
-        {
-            controller.Move(Vector3.up * Time.deltaTime);
-            if (climbingContinueTimer <= (systemTimerUpdatesPerSecond * 15))
-                climbingContinueTimer += Time.deltaTime;
-            else
-            {
-                climbingContinueTimer = 0;
-                Entity.PlayerEntity player = GameManager.Instance.PlayerEntity;
-                player.TallySkill(DFCareer.Skills.Climbing, 1);
-                int skill = player.Skills.GetLiveSkillValue(DFCareer.Skills.Climbing);
-                if (player.Race == Entity.Races.Khajiit)
-                    skill += 30;
-                Mathf.Clamp(skill, 5, 95);
-
-                if ((UnityEngine.Random.Range(1, 101) > 90)
-                    || (UnityEngine.Random.Range(1, 101) > skill))
-                {
-                    isClimbing = false;
-                }
             }
         }
